add locking to hipchat module
there is a lot of flakiness when you connect in hipchat, and it has to
do with other modules trying to send messages to the channel (even while
hipchat is still authing!).  put some manual locking around send_msg in
hipchat to keep everything settled until the bot has been configured.